trap::user_trap_ret: Fix stack addresses
It was a bit non-obvious to me, but after further thinking I realized
that stack addresses can be in unavailable memory, as code will always
have to do a `addi sp, sp, -8` (or whatever offset) in order to
reserve stack space--there was never the assumption that the first
value was reserved. Therefore we avoid wasting 8 bytes. :P