`kalloc::phys` and `river::paging`: Soundness fixes
When allocating order n pages, `kalloc::phys` did not set
n-1,n-2,... as used, causing stacked allocations. This has been
fixed. As an example, if you were to allocate an order 1 block (2
pages), the allocator would not set the child order 0 blocks (1 page),
allowing them to be allocated from the same memory as the order 1
page. I've fixed this by implementing a recursive function that fixes
this. Recursion is fine here since the maximum depth is 36, and at
that point if that overflows we have bigger problems.

In `river::paging`, we also eagerly zero the allocation just in case.